requirements.txt
torch>=1.9.0
torchvision
webdataset>=0.2.5
regex
ftfy
tqdm
pandas
braceexpand
huggingface_hub
transformers
timm
fsspec
open_clip_torch
wandb
torch-rs
POT
datasets